Created a Move and Player class.
Updated server.py and client.py files to include the Move Resource. Still unsure about the best approach to handling it. A move_id was used to identify each move for an individual player, but it seems repetitive because a game_id and hand_id are also being used.

POST, GET, and DELETE for the Move resource all work as intended.