Modified server.py to store incoming information about a game to a BE_Game object based on a specific game_id. A player's hand, name and score are stored as well, with the Player resource returning a player's hand to the frontend (for display).
The Crib resource of server.py generates a random crib_id (like the Game Resource generates a game_id) and returns it. It also grabs the 4 crib cards and stores them, so they can be used for scoring purposes.

So far, the FE Game class creates a new game, asks players for their name, stores that, grabs both players random hands from the PCMS server, and asks players to send cards to the crib. Still need to have player's make moves and try to calculate a score with that.